requirements.txt
pandas
scikit-learn
torch
transformers
pyyaml
kaggle
gradio
